Create ParticipantService
