+++ src/main/java/com/fmi/tournament/organizer/service/TournamentService.java
@@ -16,42 +16,12 @@ import java.util.UUID;
 @Service
 public class TournamentService {
     private final TournamentRepository tournamentRepository;
-    private final ParticipantRepository participantRepository;
     private final MatchRepository matchRepository;
 
     @Autowired
     public TournamentService(ParticipantRepository participantRepository, TournamentRepository tournamentRepository, MatchRepository matchRepository) {
-        this.participantRepository = participantRepository;
         this.tournamentRepository = tournamentRepository;
         this.matchRepository = matchRepository;
-    }
-
-    public void registration(UUID participantID, UUID tournamentID){
-        Tournament tournament = tournamentRepository.findById(tournamentID).orElseThrow();
-
-        if(tournament.getCapacity() <= tournament.getParticipants().size()){
-            ResponseEntity.status(HttpStatus.BAD_REQUEST).body("There is enough participants in tournament!");
-        }
-
-        Participant participant = participantRepository.findById(participantID).orElseThrow();
-        tournament.getParticipants().add(participant);
-        tournamentRepository.save(tournament);
-    }
-
-    public void unregistration(UUID participantID, UUID tournamentID){
-        Tournament tournament = tournamentRepository.findById(tournamentID).orElseThrow();
-        Participant participant = participantRepository.findById(participantID).orElseThrow();
-
-        if(!tournament.getState().equals(TournamentState.REGISTRATION)){
-            ResponseEntity.status(HttpStatus.BAD_REQUEST).body("You can only unsubscribe from the tournament while it is in the registration phase!");
-        }
-
-        if(!tournament.getParticipants().contains(participant)){
-            ResponseEntity.status(HttpStatus.BAD_REQUEST).body("There is no such participant registered in the tournament");
-        }
-
-        tournament.getParticipants().remove(participant);
-        tournamentRepository.save(tournament);
     }
 
     public List<Participant> getParticipants(UUID tournamentID){
